# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: 新点 SaaS 造价系统 API
  version: 1.0.0-current
  summary: 当前代码已注册的 V1 API 最小契约
  description: |
    自动生成文件，请勿手工编辑。
    生成命令：npm run docs:openapi-current
servers:
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ObjectEnvelope:
      type: object
      additionalProperties: true
    ListEnvelope:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: true
      additionalProperties: true
paths:
...
  /v1/projects:
    get:
      operationId: getProjects
      summary: GET /v1/projects
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: page
        in: query
        description: Page number.
        required: false
        schema:
          type: integer
      - name: pageSize
        in: query
        description: Page size.
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: postProjects
      summary: POST /v1/projects
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: page
        in: query
        description: Page number.
        required: false
        schema:
          type: integer
      - name: pageSize
        in: query
        description: Page size.
        required: false
        schema:
          type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}:
    get:
      operationId: getProjectsByProjectId
      summary: GET /v1/projects/:projectId
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/projects/{projectId}/audit-logs:
    get:
      operationId: getProjectsByProjectIdAuditLogs
      summary: GET /v1/projects/:projectId/audit-logs
      tags:
        - Audit Logs
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: resourceType
        in: query
        description: resourceType query parameter.
        required: false
        schema:
          type: string
      - name: resourceId
        in: query
        description: resourceId query parameter.
        required: false
        schema:
          type: string
      - name: resourceIdPrefix
        in: query
        description: Resource id prefix filter.
        required: false
        schema:
          type: string
      - name: action
        in: query
        description: action query parameter.
        required: false
        schema:
          type: string
      - name: operatorId
        in: query
        description: operatorId query parameter.
        required: false
        schema:
          type: string
      - name: createdFrom
        in: query
        description: Created time lower bound, ISO 8601.
        required: false
        schema:
          type: string
      - name: createdTo
        in: query
        description: Created time upper bound, ISO 8601.
        required: false
        schema:
          type: string
      - name: limit
        in: query
        description: Maximum number of records to return.
        required: false
        schema:
          type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/projects/{projectId}/default-fee-template:
    put:
      operationId: putProjectsByProjectIdDefaultFeeTemplate
      summary: PUT /v1/projects/:projectId/default-fee-template
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/projects/{projectId}/default-pricing-config:
    put:
      operationId: putProjectsByProjectIdDefaultPricingConfig
      summary: PUT /v1/projects/:projectId/default-pricing-config
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/disciplines:
    get:
      operationId: getProjectsByProjectIdDisciplines
      summary: GET /v1/projects/:projectId/disciplines
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/projects/{projectId}/members:
    get:
      operationId: getProjectsByProjectIdMembers
      summary: GET /v1/projects/:projectId/members
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putProjectsByProjectIdMembers
      summary: PUT /v1/projects/:projectId/members
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/projects/{projectId}/stages:
    get:
      operationId: getProjectsByProjectIdStages
      summary: GET /v1/projects/:projectId/stages
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: putProjectsByProjectIdStages
      summary: PUT /v1/projects/:projectId/stages
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/projects/{projectId}/workspace:
    get:
      operationId: getProjectsByProjectIdWorkspace
      summary: GET /v1/projects/:projectId/workspace
      tags:
        - Projects
      security:
        - bearerAuth: []
      parameters:
      - name: projectId
        in: path
        description: Project id.
        required: true
        schema:
          type: string
      - name: stageCode
        in: query
        description: Project stage code.
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...